move dbSchema to ns
